requirements.txt
streamlit==1.27.0
anthropic==0.7.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2
plotly==5.16.1
opencv-python==4.8.0.76
google-generativeai
qrcode
streamlit-camera-input-live